chore: add -Wthread-analysis and annotate
