+++ src/server/multi_test.cc
@@ -763,6 +763,8 @@ TEST_F(MultiTest, ScriptFlagsEmbedded) {
   EXPECT_THAT(Run({"eval", s2, "0"}), ErrArg("Invalid flag: this-is-an-error"));
 }
 
+// Flaky because of https://github.com/google/sanitizers/issues/1760
+#ifndef SANITIZERS
 TEST_F(MultiTest, UndeclaredKeyFlag) {
   if (auto mode = absl::GetFlag(FLAGS_multi_exec_mode); mode != Transaction::LOCK_AHEAD) {
     GTEST_SKIP() << "Skipped test because multi_exec_mode is not default";
@@ -794,8 +796,6 @@ TEST_F(MultiTest, UndeclaredKeyFlag) {
   EXPECT_EQ(Run({"evalsha", sha, "0"}), "works");
 }
 
-// todo: ASAN fails heres on arm
-#ifndef SANITIZERS
 TEST_F(MultiTest, ScriptBadCommand) {
   const char* s1 = "redis.call('FLUSHALL')";
   const char* s2 = "redis.call('FLUSHALL'); redis.set(KEYS[1], ARGS[1]);";
@@ -1100,6 +1100,8 @@ TEST_F(MultiEvalTest, MultiSomeEval) {
   EXPECT_THAT(brpop_resp, ArgType(RespExpr::NIL_ARRAY));
 }
 
+// Flaky because of https://github.com/google/sanitizers/issues/1760
+#ifndef SANITIZERS
 TEST_F(MultiEvalTest, ScriptSquashingUknownCmd) {
   absl::FlagSaver fs;
   absl::SetFlag(&FLAGS_lua_auto_async, true);
@@ -1118,6 +1120,7 @@ TEST_F(MultiEvalTest, ScriptSquashingUknownCmd) {
   EXPECT_THAT(Run({"EVAL", s, "1", "A"}), ErrArg("unknown command `SECOND WRONG`"));
   EXPECT_EQ(Run({"get", "A"}), "2");
 }
+#endif
 
 TEST_F(MultiEvalTest, MultiAndEval) {
   if (auto mode = absl::GetFlag(FLAGS_multi_exec_mode); mode == Transaction::NON_ATOMIC) {
