+++ src/server/search/doc_index.cc
@@ -10,6 +10,7 @@
 
 #include "base/logging.h"
 #include "core/overloaded.h"
+#include "core/search/indices.h"
 #include "server/engine_shard_set.h"
 #include "server/search/doc_accessors.h"
 #include "server/server_state.h"
@@ -17,6 +18,8 @@
 namespace dfly {
 
 using namespace std;
+using facade::ErrorReply;
+using nonstd::make_unexpected;
 
 namespace {
 
@@ -267,6 +270,20 @@ vector<absl::flat_hash_map<string, search::SortableValue>> ShardDocIndex::Search
 
 DocIndexInfo ShardDocIndex::GetInfo() const {
   return {*base_, key_index_.Size()};
+}
+
+io::Result<StringVec, ErrorReply> ShardDocIndex::GetTagVals(string_view field) const {
+  search::BaseIndex* base_index = indices_.GetIndex(field);
+  if (base_index == nullptr) {
+    return make_unexpected(ErrorReply{"-No such field"});
+  }
+
+  search::TagIndex* tag_index = dynamic_cast<search::TagIndex*>(base_index);
+  if (tag_index == nullptr) {
+    return make_unexpected(ErrorReply{"-Not a tag field"});
+  }
+
+  return tag_index->GetTerms();
 }
 
 ShardDocIndices::ShardDocIndices() : local_mr_{ServerState::tlocal()->data_heap()} {
